Add comparison functions (<, <=, radius, precedes, strictprecedes, iscommon) with tests
Corrects isthin function, so it really checks if an interval consists of
one float.

Deprecate the use of \subsetneq (in favor of \subsetdot) for interior.

Note that, by some reason, radius is not correctly exported in the tests
(but works fine in the REPL); in case of problems, it should be called
as ValidatedNumerics.radius